+++ RELEASE_NOTES.md
@@ -32,6 +32,7 @@ KeyboardKit Pro makes more locales use the iPad Pro layout, and has a new `Keybo
 * `Feedback.AudioConfiguration` has new ways to register custom feedback.
 * `Feedback.HapticConfiguration` has new ways to register custom feedback.
 * `KeyboardLayout` has new functions for using and tweaking the bottom row.
+* `KeyboardLayout` has a new functions for getting the total layout height.
 * `SystemKeyboard`'s service-based initializer now allows a layout override.
 
 ### 💡 Adjustments